saving finished / worked on webscrap / few other things

+++ main_anime.py
@@ -15,7 +15,7 @@ class Episode():
 
 class Saison():
     #classe des saison d'un animé, qui peut contenir des épisodes
-    def __init__(self,anime_id, numero_saison, nb_episodes):
+    def __init__(self,anime_id, numero_saison, nb_episodes,init = True):
         try :
             self.anime = anime_id
             self.nb_episodes = nb_episodes
@@ -24,19 +24,29 @@ class Saison():
             anime_id.dict_saisons[numero_saison] = self
             anime_id.nb_saisons += 1
             anime_id.nb_episodes += nb_episodes
-            self.init_episodes()
+            if init:
+                self.init_episodes()
+            print(f"La saison {numero_saison} de {anime_id.nom_complet} a été ajouté")
 
         except ValueError:
             print("Le numéro de la saison doit être un entier !")
 
     def init_episodes(self):
         #fonction qui crée automatiquement des épisodes avec comme nom par défaut "Episode XX"
+        # STATUS : OK
         for episode in range(1,self.nb_episodes+1):
             globals()[f"episode_{self.anime.id}_{self.saison}_{episode}"] = Episode(self.anime,self,episode,f"Episode {episode}")
 
     def nom_episode(self,numero_episode,nom):
+        # STATUS : OK
         self.dict_episodes[f"E{numero_episode}"].nom_episode = nom
         globals()[f"episode_{self.anime.id}_{self.saison}_{numero_episode}"].nom_episode = nom
+
+    def afficher_episodes_saison(self):
+        # STATUS : OK
+        list_episodes_saison = {}
+        for episode in self.dict_episodes.keys():
+            list_episodes_saison[episode] = self.dict_episodes[episode]
 
 
 class Anime():
@@ -65,37 +75,46 @@ class Anime():
 
     @classmethod
     def ajouter_anime(cls,nom_anime):
+        # STATUS : OK
         globals()[f"anime_{Anime.nb_anime}"] = Anime(nom_anime)
-        Anime.nb_anime += 1
 
     def supprimer_anime(self):
+        # STATUS : OK
+        print(f"Suppresion de l'animé {self.nom_complet}")
         del Anime.instances_anime[self.nom_complet]
         Anime.nb_anime -= 1
         print(f"L'animé {self.nom_complet} à été supprimé")
 
     @classmethod
     def afficher_anime(cls):
+        # STATUS : OK
+        list_anime = []
         print("Voici les animés ajoutés : ")
         for instance in Anime.instances_anime.values():
-            print(instance.nom_complet)
+            list_anime.append(instance.nom_complet)
+        return list_anime
 
 
-    def ajouter_saison(self,numero_saison,nb_episodes):
+    def ajouter_saison(self,numero_saison,nb_episodes,init = True):
+        # STATUS : OK
         globals()[f"saison_{self.id}_{numero_saison}"] = Saison(self,numero_saison,nb_episodes)
 
     def supprimer_saison(self,nom_saison):
+        # STATUS : OK
         self.nb_saisons -= 1
         self.nb_episodes -= self.dict_saisons[nom_saison].nb_episodes
         del self.dict_saisons[nom_saison]
-
-
+        print("La saison et ses épisodes ont été supprimés")
 
     def afficher_saisons(self):
+        # STATUS : OK
+        affich_saisons = {}
         for saison in self.dict_saisons.keys():
-            print(saison,self.dict_saisons[saison].nb_episodes)
+            affich_saisons[saison] = self.dict_saisons[saison].nb_episodes
+        return affich_saisons
 
     def webscrapp_episodes(self):
-        # STATUS : OBSOLETE
+        # STATUS : BETA
         print("Cette fonction ne fonctionne qu'avec l'animé SNK")
         print("ATTENTION : Fonctions obsolète qui ne fonctionne pas avec le reste des commandes")
         url = "https://attaque-des-titans.fandom.com/fr/wiki/Liste_des_%C3%89pisodes"
@@ -111,16 +130,19 @@ class Anime():
             print(f"Webscrapping effectué depuis {url}")
 
     def afficher_episodes(self):
-        #STATUS : BETA
-        for saison in range(0,self.nb_saisons+1):
-            for episode in range(0,self.nb_episodes+1):
-                titre_episode = self.dict_episodes.get(f"S{saison}:E{episode}")
-                if str(type(titre_episode)) == "<class 'str'>":
-                    print(f"Saison {saison} ; Épisode {episode} : {titre_episode}")
+        #STATUS : OK
+        affich_episode = {}
+        for saison in self.dict_saisons.keys():
+            id_saison = self.dict_saisons[saison]
+            for episode in id_saison.dict_episodes.keys():
+                id_episode = id_saison.dict_episodes[episode]
+                titre = id_episode.nom_episode
+                affich_episode[f"S{saison}-E{episode}"] = id_episode.nom_episode
+        return affich_episode
 
 
 def sauv_anime():
-    #STATUS : BETA
+    #STATUS : OK
     print("Enregistrement des animés en cours...")
     header = ["anime","saisons","episodes"]
     with open("anime.csv", "w", newline='', encoding ="utf-8") as main_csv:
@@ -133,7 +155,7 @@ def sauv_anime():
     print("Les animés ont été enregistrés dans le fichier 'anime.csv'.")
 
 def sauv_episodes():
-    #STATUS : BETA
+    #STATUS : OK
     print("Enregistrement des noms d'épisodes des animés...")
     for anime_to_save in Anime.instances_anime.keys():
         id_anime = Anime.instances_anime[anime_to_save]
@@ -166,7 +188,7 @@ def sauv_data():
     print("Toutes les données ont été correctement enregistrées.")
 
 def recup_data():
-    #STATUS : BETA
+    #STATUS : OK
     print("ATTENTION : La récupération des données provoque la suppression de toutes les données non enregistrées")
     print("Les animés vont être récupérés s'ils sont présents dans le fichier 'anime.csv'.")
     print("Récupération des données en cours...")
@@ -174,9 +196,9 @@ def recup_data():
         with open("anime.csv", "r", encoding ="utf-8") as main_csv:
             reader = csv.DictReader(main_csv, delimiter=',')
             for ligne in reader:
-                globals()[f"Anime_{Anime.nb_anime}"] = Anime(ligne["anime"])
-                globals()[f"Anime_{Anime.nb_anime}"].nb_saisons = int(ligne["saisons"])
-                globals()[f"Anime_{Anime.nb_anime}"].nb_episodes = int(ligne["episodes"])
+                globals()[f"anime_{Anime.nb_anime}"] = Anime(ligne["anime"])
+                globals()[f"anime_{Anime.nb_anime}"].nb_saisons = int(ligne["saisons"])
+                globals()[f"anime_{Anime.nb_anime}"].nb_episodes = int(ligne["episodes"])
         print("Les noms des épisodes des animés présents dans le fichier 'anime.csv' vont être récupérés")
         for anime_to_read in Anime.instances_anime.keys():
             id_anime = Anime.instances_anime[anime_to_read]
@@ -194,21 +216,26 @@ def recup_data():
                     globals()[f"saison_{id_anime.id}_{saison}"] = Saison(id_anime,saison,saison_found[saison])
                     id_saison = id_anime.dict_saisons[saison]
                     for ligne in reader_ep:
-                        print(ligne["saison"])
                         if ligne["saison"] == saison:
                             globals()[f"episode_{id_anime.id}_{id_saison.saison}_{ligne['episode']}"] = Episode(id_anime,id_saison,ligne["episode"],ligne["nom_episode"])
+                    print(f"La saison {saison} est ses épisodes de l'animé {anime_to_read} ont été récupérés")
     except FileNotFoundError:
         print("Aucune donnée n'est enregistrée !")
 
 
-def clean_data():
-    #STATUS : BETA
+def clean_data(type = "file"):
+    #STATUS : OK
     try:
-        print("Suppression de toutes les données enregistrées...")
-        for anime_to_delete in Anime.instances_anime.keys():
-            id_anime = Anime.instances_anime[anime_to_delete]
-            os.remove(f"{anime_to_delete}.csv")
-        os.remove("anime.csv")
-        print("Les données ont été correctement supprimées.")
+        if type == "file" or type == "all" :
+            print("Suppression de toutes les données enregistrées...")
+            for anime_to_delete in Anime.instances_anime.keys():
+                id_anime = Anime.instances_anime[anime_to_delete]
+                os.remove(f"{anime_to_delete}.csv")
+            os.remove("anime.csv")
+            print("Les données des fichiers ont été correctement supprimées.")
+        if type == "memory" or type == "all" :
+            for anime_instance in Anime.instances_anime.keys():
+                del Anime.instances_anime[anime_instance]
+            print("Les données de la mémoire ont été supprimées")
     except FileNotFoundError:
         print("Il n'y a aucune donnée à supprimer !")
